refactor(observability): split metrics tests for GH727
